feat: Add a setting which defines a custom location header (#102)
* feat: Add a setting which defines a custom location header

This adds a new setting: `location_test_header` which can be used to
specify a custom, testing location request header (e.g. `Test-Location`
which can contain a false test location for the request. This flag is
not meant for use in a production environment. If the header is not
present, the server will fall back on the normal location determination
code.

Closes #101